webadmin: Changed arrow sign from text to image
Used to have the Unicode characters for triangles as the label of a
button, but would come out garbled in non-English locales. Now using
an image instead, with the help of ShapedButton.

Also took the chance to give HorizontalSplitTable and
ImportNetworksPopupView a tighter look with the help of a designer.

Change-Id: I55f73d3edc261e4e16c19a7e345dbb1dc105c9b2
Bug-Url: https://bugzilla.redhat.com/1000285
Signed-off-by: Lior Vernia <lvernia@redhat.com>