engine: Remove instanceof from GetDomainListQuery
Runtime type checking is no way of passing parameters - if the internal
domain should not be filtered, we can just pass false.

This patch changes GetDomainListQuery to exclusively use the
GetDomainListParameters class instead of allowing
VdcQueryParameterBase to infer not filtering the internal domain.

In addition, this patch introduces a test for GetDomainListQuery to
ensure that no functionality was compromised.

Advantages of this approach:
1. Explicit is better than implicit - the code is more readable, and
   its more obvious what the caller is doing.
2. Better performance - runtime type checking is expensive, where as
   function calls are cheap.

Change-Id: I2e59b5053792e6a68030b8dcef604d9e5acdce80
Signed-off-by: Allon Mureinik <amureini@redhat.com>