core: add job that runs HA VMs which failed to run
This patch adds quartz job which is responsible to run highly available
VMs which are down. As a first step it would run HA VMs that could not
run because they were locked on last attempt.

This patch solves the following problem: when trying to run highly
available VM that went down during live storage migration, the
RunVmCommand fails because the VM is locked during the migration and the
VM won't be started again automatically. Now, when the lock could not be
acquired, the newly added job will periodically try to run the VM and
when the live storage migration will end, the VM will start.

Change-Id: I3d563d05efc6dae40f0de8e7f4b7e47aa84bd787
Bug-Url: https://bugzilla.redhat.com/891634
Signed-off-by: Arik Hadas <ahadas@redhat.com>