engine:VDS Command for Start Gluster volume remove  brick
VDS command to start the remove brick operation
on cluster volume. If force option is set to true then it
will remove the brick from volume without migrating the
data.

If force option is not present, then it first migrates all
the data from the brick being removed to other bricks. Since it is
a long runing process, gluster runs this as an async task and
returns the task ID after starting the async task.

Change-Id: I6a4d829c43a209b51d14c952bac1ada77e40083b
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>