engine: nullify EmulatedMachine field only on cluster level change
It was reset all the time because the
getParameters().getVdsGroup().getEmulatedMachine() was
always null because the FE has no notion of the emulated
machine.

The setting to null done in this patch is not really neccesary because
it is null anyway - done only to make it explicit and to be deffensive
in case the FE starts to be aware of the emulated machine in the future.

Change-Id: I8ffb180cfcd56c50b9bc228c0043e6fd5fee1c58
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1004695
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>
Signed-off-by: Roy Golan <rgolan@redhat.com>