engine:BLL Command to Start Remove Gluster volume brick
Start remove gluster volume brick is an asynchronous operation started from the console.
It starts migarting all the data in the brick(s) being removed before removing
the brick completly. On completion of the migration, user has to commit the remove
operation to finally remove the brick from volume.

Calling this command starts an async task in gluster. Gluster returns a task-id
that is used to monitor status of the task and update the UI.
StartRebalanceGlusterVolumeCommand extends from GlusterAsyncCommandBase that handles
return of task information from Gluster and updating corresponding step.

Change-Id: Ie3ee4620b75b4b714087dbf1dec3720661a5ce6b
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>