engine: Retain gluster task id on volume when task is stopped
Earlier, when a rebalance task completes, the taskid associated
with the rebalance operation that was stored in the volume
was cleared.
With this patch, we are retaining the task id as we need to
know the status of the last run task on a volume. In case of
stop, the status of task is aborted.

Change-Id: I9c5888f3b369c0f1db631085065649c2cdc75346
Signed-off-by: Sahina Bose <sabose@redhat.com>