backend: don't cache isoPrefix when empty
In rare occasions (e.g. when an iso domain was unreachable at
connectStoragePool time), the isoPrefix for the pool might be
reported empty by vdsm and eventually the value is cached (if
a VM using a cdrom is started before the iso domain becomes
available again).

This patch avoids to cache empty (faulty) isoPrefix values.

Change-Id: I6f77f92c58d9e9336e15302809195928a1ac84a6
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>