frontend: MaxVmNameLength(Non)Windows ignored
It has already been fixed but not completly correctly. There is
still a chance that the static variables will be inited sooner than
the cache from which the values are taken.

Fixed by moving the corresponding logic to AsyncDataProvider which
is anyway a better place for it and postpone the calling of it until
it is really needed so the cache will be inited for sure (otherwise
it would be not possible to reach to the code which uses it).

Change-Id: I9a1ac29aadc731f6fe4a0e62db7e78a5160ededa
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>