provider: glance images listing size and limits
This patch adds two configuration parameters to control the glance
image listing. In fact OpenStack glance servers by default return
a limited number of images (limit_param_default) and it's up to the
client to either request a longer list (using the "limit" parameter)
or to iterate with multiple requests.

GlanceImageListSize defines the number of images that should be
returned in each request (20 as the default in glance client).
GlanceImageTotalListSize defines the total number of images that
should be transferred from the server (with multiple requests if
necessary).

Change-Id: I7b534caccbdcf121b81571384e3420f1a6efde00
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>