webadmin: new VM guide dialog does not get updated
The "There are still unconfigured entities" is present
after adding a new disk.

The problem was that the postAction and resetData were
not called from the NewDiskModel.

This was partially implemented for the network interfaces
by the NewGuideVmInterfaceModel - it just was not used.

Fixed by creating the NewGuideDiskModel which works the same way as the network
one and also to use the NewGuideVmInterfaceModel.

Change-Id: I7f6c38cdc316b8f9b3501e04980d7cb2e3908136
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1005846
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>