core: revert the existing solution for bz 878041
Before having the auto-start VMs runner job, the problem of HA VMs that
could not be restarted when failing during live snapshot operation was
solved by comparing the state of the VM at the end of the live snapshot
operation with its state when the command started, and if we saw that
the VM went down unintentionally we ran it at the end of the live
snapshot command.

That solution was wrong since it was specific to live snapshot operation
and it required to put large portion of the code in endVmCommand method
inside new transaction scope. Now that we lock the VM for the entire
live snapshot operation and we have the auto-start VMs runner job that
keep trying to restart failed HA VM while it is locked, which is a more
general and clean solution that solves this case as well, we can revert
the previous solution.

Change-Id: I242a2278c9c23139026b58564ee35967e5258bfe
Signed-off-by: Arik Hadas <ahadas@redhat.com>