core: lock VM for whole live snapshot process
VM was used to be locked for the entire live snapshot process only in
case the snapshot contained memory. If the snapshot did not contain
memory, the lock was released in the execute state of the command.

This patch changes that, such that the VM will be locked for the entire
live snapshot process in any case. By keeping the VM locked for the
entire process, we prevent other operations which should not run in
parallel to run, as long as they also try to lock the VM.

Change-Id: Ib455a3b27d7b032c6a706753a6e58e07d3a1bc14
Signed-off-by: Arik Hadas <ahadas@redhat.com>