webadmin: remove bricks with migration support
Added a new Checkbox 'Migrate Data' to the Remove Bricks
Confirmation popup. And a warning message in the bottom
explaining the necessity to migrate the data before remove.

The bricks in the confirmation popup is wrapped inside
a ScrollPanel to avoid the overflowing if more number of
bricks are selected.

Change-Id: I8aae4548661ff1cb9f93a14ab66b6c41bc201bbd
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>