core: Avoid XSS in RedirectServlet
Currently the RedirectServlet composes JavaScript code to show error
messages using text provided by the user in a request parameter. This
text isn't sanitized and thus can be used by maliciuous users to execute
arbitrary JavaScript code. To avoid this situation this patch changes
the servlet so that it doesn't receive any parameter, thus the problem
is completely avoided.

Signed-off-by: Alexander Wels <awels@redhat.com>
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Change-Id: Ie77e6a063e1522b2e108076a240939ca1dae272e