engine: Engine has to delete image from DB after VDSM removed it
Engine has to delete image from DB after VDSM removed it from storage,
but thrown an error in the process. What is really matter in this case
is whether the image still exists on the storage or not.
VDSM renames the image as a first step to removal. After renaming the
image doesn't exist, so if the error is thrown at this point the garbage
will stay on the storage, but its safe to remove image from DB.

Change-Id: If6e6922b153145de6d4515812c1cfede687544bc
Signed-off-by: Sergey Gotliv <sgotliv@redhat.com>