webadmin: Attempt to add a 3.2 host into 3.3 env fails but the host cannot be removed
When a 3.2 host is added to 3.3 env the server returns
NonOperationalReason___CLUSTER_VERSION_INCOMPATIBLE_WITH_CLUSTER
message. Since this message is not available in the frontend
there is an exception thrown when a user clicks on the
row for the host. So the row is not selected
resulting in the action buttons not being activated.

This patch adds a new message to LocalizedEnums

Change-Id: I407154df67d51f28117f7aaf51f8030e39934678
Bug-Url: https://bugzilla.redhat.com/1004675
Signed-off-by: Ravi Nori <rnori@redhat.com>