restapi: Separate User and DomainUser user resources
Currently we have two conceptual resources to represent users:
UserResource and DomainUserResource. These are separated interfaces, the
first representing an user that has been added to the database of the
engine and the second representing an user that exists in an external
directory and that may or may not have been added to the database. These
two different conceptual resources are implemented then using a base
class AbstractBackendUserResource. The problem is that due to the
generic nature of this base class both implementations are forced to
specify DbUser as its type, instead of DbUser for the first concept and
LdapUser for the second concept. This complicates changes to DbUser and
LdapUser. To make that easier this patch removes the abstract base class
and makes both implementations use the generic type that is natural for
them.

Change-Id: I428f98772ac57cd7825781fb3cf282b2ac56e995
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>