webadmin: Replaced null profile with empty profile
Added a special value to represent an empty profile, and changed empty
profile instances to that instead of null. This is needed by a
soon-to-be-implemented Add/Remove VNIC widget in the Add/Edit VM
dialog, in order to discriminate between the case where a profile
hasn't yet been assigned (null) and the case where the empty profile
has been assigned (VnicProfileView.EMPTY).

Change-Id: I06962e43084ff4fd469e0025527cdd85a47f5ea7
Signed-off-by: Lior Vernia <lvernia@redhat.com>