webadmin: volume rebalance actions on the row
Moved volume rebalance actions to the individual row.

Added a new clickable 'down arrow' icon to the volume
activity column. On clicking the 'down arrow' icon,
rebalance actions will be shown in a menu.

A seperator is shown between the rebalance status icon and
the 'down arrow' menu icon.

New Composite cell is introduced to contain rebalance status icon,
seperator image and the menu icon. This cell will be visible
only if there is task associated with the volume.

Removed stop and status rebalance actions from the action bar.

Change-Id: I49331d1b21cd43a9c2b9b0251ceb34fdb016aaf8
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>