core: increase AutoStartVmsRunner job frequency
This patch changes the default frequency of the AutoStartVmsRunner from
60 sec to 1 sec. Instead of invoking the RunVmCommand on each iteration,
the AutoStartVmsRunner now try to acquire the needed VM lock first, and
only if it succeed it invokes the RunVmCommand and the command release
the acquired lock.

The type of the internal collection of HA VMs to run in
AutoStartVmsRunner job is also changed to CopyOnWriteArraySet which is a
more appropriate data structure for this use-case and for the following
planned enhancements.

Change-Id: I0968cae554458f53a1f30643848bb8d88bd4de11
Signed-off-by: Arik Hadas <ahadas@redhat.com>