core: change the retry to run failed HA VM frequency
The AutoStartVmsRunner default frequency is 1 sec, in each iteration it
iterated the down HA VMs in his set, tried to acquire a proper locks
for running them and then invokes the RunVmCommand for each of them.

The behavior above is ok for the scenario where you want to try to run
the VM after you failed to do so because you did not manage to acquire
the lock (because some other command(s) took the lock and run in the
background) - you want to keep trying in high frequency in order to run
it when the operation(s) in the background is finished.

But if the VM cannot be run from other reason, because of scheduling
constraints for example, trying after 1 sec could make bad effects and
is not likely to succeed. So in that case we will wait more before
trying to run the VM again.

New configuration value is added to define the interval (in seconds) we
will wait before trying to run the down HA Vm after not being able to
run it not because of lock acquisition problem.

In addition, we limit the number of times we try to run failed HA VM
(not because of lock acquisition problem). New configuration value was
added for the maximum number of tries, and if we exceed this number of
tries, we generate proper audit log message and we don't try to run it
automatically anymore.

Change-Id: I9f8914efc73ac1ea4f4a2ec4e588d778e9002451
Signed-off-by: Arik Hadas <ahadas@redhat.com>