core: run HA VMs that went down right before engine stopped
This patch solves a known problem that was exists before: when HA VMs
went down, the engine (VdsUpdateRunTimeInfo) detected it (and updated
the DB) but didn't manage to run the VM, we won't try to run the VM when
the engine start.

This problem is solved by fetching from the DB all the HA VMs which are
down with exit status ERROR when initializing the AutoStartVmsRunner
job, and those VMs are set as the first VMs the job will try to run on
its first iteration.

Change-Id: I2d5876f196819b2a69be0b71287c5325a8ff9dc9
Signed-off-by: Arik Hadas <ahadas@redhat.com>