core: add VMs from the same VDS to auto startup job in bulk
VdsUpdateRunTimeInfo checks all the VMs which run on the VDS it
monitors, and at the end of the process it has all the HA VMs that ran
on that VDS and went down. So Instead of iterating over those VMs and
pass each of the separately to the AutoStartVmsRunner job, it now pass
them as a bulk (through the VdsEventListener) which is a bit more
efficient.

Change-Id: I560383c8e73bb1728441fe5786954eba7ba88e60
Signed-off-by: Arik Hadas <ahadas@redhat.com>