core: use AutoStartVmsRunner job from VdsUpdateRunTimeInfo
Currently VdsUpdateRunTimeInfo invokes RunVmCommand for each HA VM that
went down and only when it fails because of lock acquisition problem,
the VM is added to the AutoStartVmsRunner job.

This patch changes that, such that the VdsUpdateRunTimeInfo pass the VM
to the AutoStartVmsRunner right away and it will try to run the VM in
its next iteration.

Change-Id: Ia5ff535810f969255fb1d496a11378c2c10bcb91
Signed-off-by: Arik Hadas <ahadas@redhat.com>