engine: Set correct size of iso files and floppies
Previously all iso and floppy files appeared in UI with bogus size
of '< 1 GB' although some of them are larger than that. Furthermore it
would be nice to display file size in human readable way.

The root cause of that problem is 'getIsoList' API between Engine and
VDSM which returned only a list of file names until now, so Engine set
their size to 0. It's worth mentioning before explaining API changes
that this list contained only permitted files.

Now this API returns a map with file stats including a size for all
files including not permitted files. Meantime Engine filtering out
files with wrong permissions, but they can be displayed in the future so
user would be able to know why these files are not usable.

Engine uses -1 when size is not available because old versions of VDSM
don't send it.

Change-Id: I77bd99beb8138524b25f0afdcce0815ad8664f0f
Bug-Url: https://bugzilla.redhat.com/1005889
Signed-off-by: Sergey Gotliv <sgotliv@redhat.com>