engine,frontend: localize snapshot type on FE
The problem was that the "Activet VM before preview" and "Active VM" Strings
were hardcoded in the backend and written to the DB as a description. The
problem with this was that the strings were taken on the frontend "as is"
and were not localized.

This patch localizes this strings on FE but leaves tham on BE to provide them
to REST.

Change-Id: I26866fd4dd91d034e7ae5afee054080a4ab864a1
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1010122
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>