webadmin: Attempt to add a 3.2 host into 3.3 env fails but the host cannot be removed
When a 3.2 host is added to 3.3 env the server returns
NonOperationalReason___CLUSTER_VERSION_INCOMPATIBLE_WITH_CLUSTER
message. Since this message is not available in the frontend
there is an exception thrown when a user clicks on the
row for the host. So the row is not selected
resulting in the action buttons not being activated.

This patch adds a new message to LocalizedEnums

Change-Id: Ib9763a444cb4c96d407e7e8bee55b72bccaff89f
Bug-Url: https://bugzilla.redhat.com/1004675
Signed-off-by: Ravi Nori <rnori@redhat.com>
Signed-off-by: Tal Nisan <tnisan@redhat.com>