core: Prevent NPE with cobertura
The NetworkDaoDbFacadeImpl contains an abstract row mapper and then two
derived ones, and all of them, including the abstract one contain a
singleton. The singleton in the abstract one is never used, and it
causes a NPE if used. Some tools, like cobertura, generate code that
uses all the fields, and thus reveal this hidden NPE. This patch just
removes the unused field.

Change-Id: I811fe567010037fd12e4deeb517fb3396c724b53
Bug-Url: https://bugzilla.redhat.com/1012455
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>