webadmin: Disable external networks in Setup Networks
Changed the appearance of external networks in Setup Host Networks to
look disabled and not be draggable, as they truly shouldn't be
statically assigned to the host's interfaces.

Did this by extending the hierarchy of related
classes. InternalNetworkPanel should behave like a native network,
while ExternalNetworkPanel is a disabled variation of it. Common code
is housed in the common ancestor NetworkPanel. Also refactored some
code around UnassignedNetworksPanel for reuse.

Change-Id: I556ec7a330236dbdb204b82b2ff67f86ed4dbea8
Bug-Url: https://bugzilla.redhat.com/990612
Signed-off-by: Lior Vernia <lvernia@redhat.com>