packaging: setup: partial cleanup - db only
Before this change, if during cleanup the user only asked to delete
only the database, then ran setup again, setup wrongly identified
the system as being legacy 3.2 and tried to upgrade, which failed.

This change makes the identification of upgrade from 3.2 legacy
more strict and in such a situation setup-engine will continue as
a new setup.

Bug-Url: https://bugzilla.redhat.com/987939
Change-Id: Id9aa1791eb7ccdbc8c679e72ed6eaa4f72ac29c6
Signed-off-by: Yedidyah Bar David <didi@redhat.com>