webadmin: Allowed non-draggable NetworkItemPanel
It seems that so far setting draggable to false did nothing, the item
would still be draggable; for some reason calling
setDraggable(Element.DRAGGABLE_FALSE) adds the draggable property to
an element and sets it to true!

Added some code to disable the "draggable skin" and the actual drag
operation when an item has been constructed as non-draggable. Also,
changed bonds to be draggable explicitly, as they have been draggable
all this time and I've heard no complaints.

Change-Id: I0f130803ce2403cc26297ec8d06ef62d23837b96
Signed-off-by: Lior Vernia <lvernia@redhat.com>