restapi,engine: Support network name in update nic
The network name and the port mirroring flag provided
by the restapi for updating a vnic name should be
respected. In this case the vnic profile id will be ignored.
In addition, the patch allows to set 'null' value for the
vnic profile id, which is treated as unlinking the nic from
the network:

<nic>
  <vnic_profile/>
</nic>

This patch designed for backward compatibility only of the
add/update vm nic.

Change-Id: I1f5c96daf64f6e14cc76bfb12daf89618f656b3d
Bug-Url: https://bugzilla.redhat.com/1011116
Signed-off-by: Moti Asayag <masayag@redhat.com>