core: remove memory volumes without tasks on import vm failure
On failure to import vm with snapshots that contain memory state, we
remove the imported memory volumes.

This patch fix a bug which is relevant for the memory volumes removal in
this scenario, that was caused by not adding the tasks properly (not
override the getTaskType method in CommandBase). The solution is to
remove those imported memory volumes without tasks, in a similar way to
the solution done for I3830992f46b2c3fbea47e69f21a6711718408e95.

Change-Id: I596ea3164c70b6c2a4e2c161737421259457a179
Bug-Url: https://bugzilla.redhat.com/1015080
Signed-off-by: Arik Hadas <ahadas@redhat.com>