core: Use pin to host only when present in cluster policy
This removes the explicit pin to host check in scheduling.
The check is already present in cluster policy and will be used
when the policy unit is selected by the administrator.

It also restores the optimization we had in the code that
used Pin To Host logic as the first filter for the provided
cluster policies.

Change-Id: If2c0496327c9f692b0dbe92e1cf481f900f4a747
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1007282
Signed-off-by: Martin Sivak <msivak@redhat.com>