restapi: VmMapper test - Move mock to class level
When mapping OsType, the code references a service which is currently mocked.
But it's not mocked everywhere it should be, and the result is a strange
behavior: when all the test method in VmMapperTest are run sequentially,
one after the other - they pass. But if only 'testRoundTrip()' is run - it
fails.

The current change takes the necessary mocking up to the class level. The mocking
is static and should only happen once - that's why the class level is the
right place. And this way all test methods benefit from the mocking.

Change-Id: Iaa9794a268551461263ff0aecb12a12cda279bd1
Signed-off-by: Ori Liel <oliel@redhat.com>