core: Fix PinToHostPolicyUnit when PinToHost is enabled
We have to return an empty list of available hosts when PinToHost
is enabled and the pinned Host was already filtered out.

Otherwise the whole PolicyUnit has no effect.

This is fixing one issue out of two that caused the
referenced bug. The issues are independent on each other
and so I did want to preserve the history.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1007282

Change-Id: Id5a49813d433a10debe607f17783057d7d4a767b
Signed-off-by: Martin Sivak <msivak@redhat.com>