webadmin: Added vertical scrolling to BaseListModelSuggestBox
All suggest boxes in the system could probably use a height limitation
for their suggestion list, with a vertical scrollbar when the content
is too big.

Also took the chance to move the logic fitting the height of the
suggestion list to that of the content from ListModelTypeAheadListBox
to BaseListModelSuggestBox, as ListModelSuggestBox could also use it.

Change-Id: I9c662e3e8d1062e654f10ce275019c28b62f2666
Bug-Url: https://bugzilla.redhat.com/1008523
Signed-off-by: Lior Vernia <lvernia@redhat.com>