webadmin: Open correct bond dialog from Network/Hosts subtab
Apparently the old (pre 3.1) bond dialog would open, when the newer
one should. Also removed code that checked if the open dialog is the
Setup Networks dialog, which as far as I know is the only possible
case in this subtab (whereas in Host/Networks it could have been the
pre-3.1 Host Interfaces dialog).

Change-Id: Ia5cc40bda079513bfce7b08d6f279aaebe554817
Bug-Url: https://bugzilla.redhat.com/1012821
Signed-off-by: Lior Vernia <lvernia@redhat.com>