frontend: Allow using any domain for sysprep
This patch allows user to alter domain name for sysprep when running the
machine using "run once".

The problem was caused by wrong condition when filling from model to
RunVmOnceParams. Fixed by adjusting the condition.

There was also another issue with the ComboBox, which wasn't working
correctly when there is a custom text put in and user wants to apply a
value from the drop-down menu. In this case the value is not updated.
Hot-fixed by manual handling of DOM events.

Change-Id: Iff87239de98694cf124a44e7cb39f6d9e867ce93
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1015154