gluster: Removed check of async task details in volume
Removed checking the existence of async task details in volume entity
for the purpose of checking the rebalance execution status on a volume.
This is not required and we should always depend upon the actual value
returned from glusterfs for the rebalance status as that value would be
the uptodate and correct value.

Change-Id: I54a0f0949cfc8c2fb46bec816519449e583a6b65
Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>