engine: Use getFileStats to retrieve iso and floppy files
VDSM introduced new API getFileStats in 3.4 which returns statistics
of all files in the domain filtered according to pattern.

Now Engine starts using this new API, because file statistics are
significantly better in comparison with the existing getIsoList and
getFileList and allow Engine to provide additional information in UI,
for example a file size.

Change-Id: I735d2c9343fe791e0890b4a6cc9f12c74462b3f0
Bug-Url: https://bugzilla.redhat.com/1005889
Signed-off-by: Sergey Gotliv <sgotliv@redhat.com>