webadmin: refactored volume activity status column
Removed the hard coded GlusterVolumeEntity from all the
volume activity related cells and used GlusterTaskSupport instead.
So that the columns/cells can be re-used for showing the
activities related bricks as well.

GlusterVolumeEntity and GlusterBrickEntity is inherited
from GlusterTaskSupport.

Used a diffrent icon for the menu cell to have better visibility.
Also the activity cell will be available if the task status is not-null
and not UNKNOWN.

Change-Id: I85df6e5d2071f8ff1cc5f94ace90edfb0fff471e
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>