backend, restapi: Add read-only disk functionality
All underlying components are already supporting RO disks
functionlity. After that change Engine will support it either through
the REST API.

When adding/attaching/editing a disk to a VM, this feature adds a
read-only property to the VM-Disk relationship which is persisted
through VMDevice (vm_device in the DB). Floating disks continue
to be RW.

Note that shareable disk could be attached to one VM as RO, and to
another as RW.

Change-Id: I124176e8feb91b601a71e76dd63051648ec4353a
Bug-Url: https://bugzilla.redhat.com/878705
Feature-Page-Url: http://www.ovirt.org/Features/Read_Only_Disk
Signed-off-by: Vered Volansky <vvolansk@redhat.com>
Signed-off-by: Sergey Gotliv <sgotliv@redhat.com>