webadmin: Don't display DCs that imported external network
Changed the import networks dialog so that the choice of DC to import
an external network to, will only include DCs where the network
doesn't exist yet.

This required modification of the backend query to not only retrieve
all the external networks that exist on a provider, but also map each
network to the set of DCs where that network had already been
imported. This set is then used to efficiently check whether a certain
external network had been imported to a certain DC.

Change-Id: Ifd76b527500957ec47db6d60894b0d7316d17058
Bug-Url: https://bugzilla.redhat.com/989909
Signed-off-by: Lior Vernia <lvernia@redhat.com>