engine: quota was not enforced in run VM
Since quota uses MLA model, each quota related action should
add permission checks for Consume_Quota action group on the
specified quota. this was not the case in Add VM and the command's
derivatives (add VM from .., etc.).

The reason for that was that the user can create a VM from template
and select the template's quota. this quota should be available only if
the user has consumption rights for this quota.

Last issue is that when failing on no-permission of consume_quota, the
err message shouldn't be: 'the user isn't author...' but more suitable
one related to quota and consumption.

Bug-Url: https://bugzilla.redhat.com/1020201
Change-Id: I053680e61b8f9a0557845d0b5359a022b6473052
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>