webadmin: adding activities colume to Bricks tab
Added a new activities column to Bricks table.
This is similar to the Actvities colume in the volumes tab.
But the cell in this column is visible only if the task
is of type REMOVE_BRICK. Added new icons to show the
different stages of remove brick operation.

All the brick removal related(stop/commit) operations are moved to
the row. The same actions will be available in the volume row as well.
Stop/Commit removal command availbility will be decided based on
the task type and status.

Refactored the code which populates the bricks in the table.

Change-Id: I820c8b2e2e948d1c2ce0811d2d5a78784eeafc2b
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>