engine: Move quota to Image_storage_domain_map
In order to allow assigning different quota to the same image on
different storage domains, the quota id was moved from the images table
to Image_storage_domain_map table.
Backend entities were change to suit this change - so that quota is no
longer a field of Image but a field DiskImage and multiple quotas can be
assign to the same DiskImage for different Storage Domains.

Bug-Url: https://bugzilla.redhat.com/1001032
Change-Id: Icc81b077cf125de85fcb3586bc0d0f71e5e0716a
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>