webadmin: adding gluster volume remove brick status
Added a new 'Status' actions to the remove-brick menu.
On clicking status action, 'Remove Brick Status' popup will be shown.

remove-brick uses rebalance internally for the data migration.
The status output for remove-brick will be same as rebalance.
So re-used all the ui components from the rebalance status.

'Stop' and 'Commit' remove-brick action are added
to the status popup as well.

Change-Id: Iaed9cfeb52e990d2059b628598e1b82daf1d01e5
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>