restapi: Add support for JAXB annotations in JSON
The default JSON provider included with RESTEasy 2.x doesn't take into
account the JAXB annotations, in particular it doesn't take into account
the annotations that specify the names of elements and this generates
inconsistencies in the names used in the XML representation and in the
JSON representation. In order to solve the problem this patch modifies
the JSON provider used by the application so that it takes into account
the JAXB annotations.

Change-Id: Ica39bbedd427dbf2f298b1292fe9a14f3825d6e5
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>