engine: enabling StopRemoveBrick when data migration is completed
Changei canDoAction to allow Stop Remove brick to be called even when the
task is finished( data migration complete). This helps to implement
the Retain Bricks UI command. Also change the remove-brick step status
to ABORTED.

changing the step status from FINISHED to ABORTED is not allowed through
Step.markStepEnded. Hence explicitly update the status and endtime in
GlusterAsyncCommand.endStepJob.

Change-Id: I88e53fa1772427ea1ae2b6943b76e9a71e772376
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>