core: Clean pending memory and CPUs only during init
When the engine tries to start a VM, we leave pending memory
and CPU information in the database. This information is removed
when the VDSM reports the VM as fully Up.

However when the engine crashes without telling VDSM, the info
needs to be removed as no VM was actually started.

We dealt with this in VdsUpdateRuntimeInfo so far, but that caused
a race condition: We removed the pending data because the engine
did not manage to send the VM start event to VDSM before the next
VDSM update.

This patch moves the cleanup to the engine startup routines to
make sure it gets applied only in cases of real engine crashes.

Change-Id: I84c2ba29368c0b6bc400087512d7b3a68c5c8f5c
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1008146
Signed-off-by: Martin Sivak <msivak@redhat.com>