core: Remove non-existent Storage Domain will be rolled forward.
Removal of a Storage Domain which no longer exists causes a failure in
the engine and blocks the user from removing the storage.

The following patch validates the storage domain was not removed due
to the reason of non existence, and if so it will roll-forward and delete
it any how.

Change-Id: I9515a5ddfcaa965c2bad1883f0543494915b8c7d
Bug-Url: https://bugzilla.redhat.com/1001626
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>