frontend: packaging: Make default VNC console mode configurable
This patch adds an (engine-wide) option to configure default VNC client
console mode (both SPICE and RDP have this option already). Currently
admin can set Native or NoVnc client modes.

If NoVnc is chosen and is not available on the engine (WebSocketProxy
config not set), Native is used as a fallback.

Change-Id: Icb37fe653f007307686765f6dc089206559cb7e3
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1035719