frontend: NPE in Import*Model
There was a NPE vulnerability in ImportTemplateModel and ImportVmModel.

The problem was that the VmBackupModel.restore() called first the
setItems() and than init(). The setItems have done some async call
and set the result to the super.setItems(), while the init() already
needed this in initDisksStorageDomainsList() method (which is also
called in a callback). So if the callback from setItems returned sooner
than the ones in init() it worked. If not, the initDisksStorageDomainsList
failed on NPE.

Fixed by calling the real init() logic only when the data are really prepared.

Change-Id: Id97c8368ecf80ceb7ef1aa456d32a022dd5043bb
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1019862
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>